+++ finagle-core/src/main/scala/com/twitter/finagle/loadbalancer/p2c/P2CLeastLoaded.scala
@@ -6,7 +6,7 @@ import com.twitter.finagle.loadbalancer.LeastLoaded
 import com.twitter.finagle.loadbalancer.Updating
 import com.twitter.finagle.NoBrokersAvailableException
 import com.twitter.finagle.ServiceFactoryProxy
-import com.twitter.finagle.loadbalancer.LoadBalancerFactory.PanicMode
+import com.twitter.finagle.loadbalancer.PanicMode
 import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.util.Rng
 import com.twitter.util.Activity